requirements.txt
altgraph==0.17.4
Babel==2.14.0
docxcompose==1.4.0
lxml==5.1.0
packaging==23.2
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
pypandoc_binary==1.13
PySide6==6.6.2
PySide6_Addons==6.6.2
PySide6_Essentials==6.6.2
python-docx==1.1.0
shiboken6==6.6.2
six==1.16.0
typing_extensions==4.10.0
